requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
cryptography==39.0.1
defusedxml==0.7.1
Django==4.0
django-countries==7.5.1
django-filter==22.1
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.13.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.25.1
exceptiongroup==1.1.0
flake8==6.0.0
flake8-import-order==0.18.2
flake8-isort==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
loguru==0.6.0
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.0
pluggy==1.0.0
prompt-toolkit==3.0.37
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.1
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
